[add]AgentReader追加

+++ src/main/java/com/roborescuemod/commons/AgentData.java
@@ -10,7 +10,7 @@ public class AgentData {
 
 	public Entity agent = null;
 	public World world = null;
-	public ArrayList<Integer> path = new ArrayList<>();
+	public ArrayList<Double> path = new ArrayList<>();
 
 	public AgentData(World world, Entity agent, Point3D point) {
 		this.agent = agent;
@@ -22,22 +22,22 @@ public class AgentData {
 		agent.move(MoverType.SELF, x, 0, z);
 	}
 
-	private int pop() {
+	private double pop() {
 		if (path.size() != 0) {
-			int p = path.get(0);
+			double p = path.get(0);
 			path.remove(0);
 			return p;
 		}
 		return 0;
 	}
 
-	public void setPath(ArrayList<Integer> history) {
+	public void setPath(ArrayList<Double> history) {
 		path.addAll(history);
 	}
 
 	public void shift() {
-		int x = pop();
-		int z = pop();
+		double x = pop();
+		double z = pop();
 		this.move(x, z);
 	}
